requirements.txt
SQLAlchemy>=2.0
requests>=2.31
python-dotenv>=1.0
matplotlib>=3.7
termcolor>=2.3
